Refactored MainActivity by extracting the Compose application root and onboarding flow into a new BitchatApp composable. This reduces the activity's responsibilities to primarily handling Android lifecycle, permissions, and service coordination, while delegating the UI tree to a dedicated top-level composable.
